Add insert(path, value) and put(path, value) methods to Value.
These mutators assume the Value is an Object and use a path notation
to insert values. They traverse the object hierarchy, creating
additional parent objects as necessary, then setting the value. insert
follows standard container semantics of not overwriting existing
values, put does overwrite values. The separator is a parameter, but
defaults to '.'. Also add tests for this functionality.