Add PathError and use it when bad paths are encountered.
Multiple kinds of bad paths could be used on insert(), put(), or, in
the future, on gets(): called on non-objects, using an empty subpath
(i.e. two delimiters in a row), hitting components in the path that
aren't actually objects, etc. This adds the exception, uses it for
insert() and put(), and adds tests for the functionality.