add: Notification panel, improve: Application status icon
The notification panel and the StatusItem are defined as a part of the
FinderWindow framework, but instanciated in the main application (in the
contrary of all other parts injected into the Finder).

The notification panel is designed as an highly customizable panel view,
while the status item is re-implemented in order to intercept gracefully
all mouse events, and more importantly, to know its exact position in
the main status bar.

A lot of this was taken from this demo: https://github.com/shpakovski/Popup.git
(which is BSD licensed)