improve: FinderWindow and InfinitApplication relationship.
Some additional components provides simple ways to use gap even when
the main infinit application (which is responsible for notifications,
main sidebar and main menu) is in a different process than the infinit
windows (main application functionalities). Gap is now instanciated
twice, one in the finder windows, and another in the main Infinit
application with the token resulting from meta authentication.

The notification panel is now designed and defined in the main app,
while other components stay in the FinderWindow.