aws.S3: Reorganize exception, hook on transient errors.
AWSError now contains a sub-exception with the lower-level problem,
and additional data.

A simple hook on_error is called with an AWSException on both transient
(internal retry) and fatal (exit with exception) errors.

AWSException::operation is just the url for now, to be augmented with
specifics in a later patch.