aws.s3: Fix double-quoting of etag in multipart_finalize.
Somehow it did not cause any problem all(most?) of the time, but we have
rare "malformed xml" errors from S3 in the logs.

Wild wild guess: aws is very leniant in what they parse, and only fail on
malformed etag if there are multiple possible etags value for one part,
which occurs in case of part resend.