Fix MSVC shared lib build
Define `BOOST_ALL_DYN_LINK` when building as a shared library using
Visual Studio. Without this definition, the Boost auto-linking feature
links to the Boost static libraries.

Set `NETWORK_URI_DECL` to `BOOST_SYMBOL_IMPORT` when using the
library instead of building it. This definition expands to
`__dllspec(dllimport)`, which is required to use the shared library in
Visual Studio.