`NETWORK_URI_DECL` per function, not per class
This commit removes the `NETWORK_URI_DECL` declaration from all classes
using it and instead applies the declaration to individual functions and
member functions that need it.

The problem with applying the declaration per class is that it causes
compiler warnings when using Visual Studio--both when building the
library and when using the library from another project.