Buffer: factor and address a latent bug
In one of the ctors of InputBufferArchive, my compiler was complaining
about _istream being used before it's initialized.  That's because we
want to pass to the super class's constructor (!).

Apply the same trick as in the other ctors, but rather than
duplicating this trick, factor it using an additional constructor and
delegate to it.

* elle/src/elle/Buffer.cc, elle/src/elle/Buffer.hxx
(InputBufferArchive, OutputBufferArchive): Enforce order of initialization
of _istream/_ostream.