protocol: let the Streams know the protocol version
Now that the coding of the integers depends on the protocol version,
Stream and its descendants need to know the protocol version, which is
known by their backend.

Let them expose a version() function that returns that of their
backend.  So ChanneledStream no longer need to aggregate an
elle::Version, they just have to query their backend.