serialize: use a virtual function to tell the difference between in and out
Previously out() was running a dynamic_cast to tell the difference
between SerializerIn and SerializerOut.  For performance reasons, it
was changed in ab5174f107bc250d9569364f6fc9fb39c66e4468 to use a
Boolean member.  I believe that using a virtual function should be
about the same, yet cleaner.