Update the mtime of symlinks, not their destination
We use lstat to get the mtime of a file.  That's on purpose, since we
also build symlinks.  But then, when we need to update a file that is
a symlink, we must update the source, not the destination.

This fixes the problem that sometimes we declare that we "Failed to
adjust mtime" of files such as libboost_context.a.  That's because
it's actually a symlink to libboost_context-1_60.a, and we keep
updating the destination and check the source.