elle.printf: reduce code bloat
Use unnamed namespace for fprintf, sprintf, and printf, to allow the
compiler not to leave these specializations in the symbol table: they
won't be used elsewhere, and if they are, it will make other copies.

Introduce format_error to factor the generation of the error messages.
Make it non-template.  And call it from elle::format rather than from
its callers, as that's really where the error can be raised.