elle.athena.paxos: fix indentation and space issues
The previous commits were made before the change of namespace (athena
becoming elle::athena), and applied after.  As a consequence, the
indentation was wrong.

Most of the work here was made by clang-format.  But I had to
cherry-pick the lines, as clang-format does not allow to do everything
we want.  For instance, always breaking before the list of base
classes.