build: fix an incorrect lib_path_runtime
libcrypto had '../../../lib' as rpath, instead of '.'.  As a result
the BF was using ancient shared objects installed long ago, instead of
the current ones.  So it broke when a symbol was introduced for the
first time in a long time.

Now the question is: why is the Mac the only machine to have a problem
here...