elle.reactor: help clang accept for-each
For clang, the decltype of a member of a const structure does not have
`const`.  As a consequence, the decltype of captures of non-mutable
lambdas are "not visibly const" but they are.  Add the missing const
for clang.

GCC does make the decltype of captures `const`, although it does not
do it for structs.