drake: fix the handling of reactor examples
We have two different conventions for examples: until recently all of
them where handled by hand, relatively to the examples/ directory.
Recently, echo-server was moved to be completely handled by drake, and
therefore relatively to the root directory.

As a result, elsewhere the code adds a duplicate examples/ prefix to
the examples of reactor.

We should not have two conventions.