elle.log: Windows portability
The tests were creating several ofstreams, and kept them open.  But
the clean up procedure was removing the directory that contained them,
which does not work on Windows (ends with a "sharing problem").

So we have to keep track of the open ofstream and close them when we
are done with them.  That's exactly the purpose of FileLogger.  Hence,
generalize it to support the possibility _not_ to rotate, and to
append.