elle.log: use std::chrono
This kind of reverts 79d634ea261ec5d278037f3150d35091869100dc by
storing the interface based on Message: now the Messages cary their
time stamp, and it is only when printed by a logger that the format of
the conversion to string is applied.

There is one problem though: currently we don't support local time
display, only universal time.