elle.Backtrace: clean up the CPP symbols
We have a lot of duplication on the logic for NO_EXECINFO, and then to
know whether we do support backtraces.

Let's factor these under the name ELLE_HAVE_EXECINFO and
ELLE_HAVE_BACKTRACE.  And let's get rid of commented code.  And let's
have StackFrame have a ctor that hides part of the backtrace_symbols
decoding mess.

This is prompted by the fact that this test fails on Alpine where we
do not have backtrace support.