elle.bench: be parameterized by different types
Many of our uses of Bench are based on adding doubles to sum number of
milliseconds.  We should never do that, and always use some
std::duration (such as elle::Duration).  Yet there are places where
Bench is used to monitor a number of calls.

So make it parameterized by what we measure.